modemloggerd: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=modemloggerd \
        -n 1 \
        --platform2-commit=ebb8e32e609ba82b3bd540db54baa7c937216fb7

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to login_manager and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: I300d46de4f8557917a83ba13ef773c2ba9edaa5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5950727
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>