patchmaker: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=patchmaker \
        -n 1 \
        --platform2-commit=930387fce539ad73aca30e3b8362d80891adeb12

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to mtpd and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: I325131aa0bbdac1a435a3b1df844a785006e0598
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5950331
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>