pciguard: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=pciguard \
        -n 1 \
        --platform2-commit=930387fce539ad73aca30e3b8362d80891adeb12

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to mtpd and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: Ic39678935e05929fbc2ce764cf6149caad3c8214
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5950333
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>