arc: Add Verify and Copy Device IDs
For Android runtime on ChromeOS attestation project,
we need to attest various identification labels
of the device.

This code change list deals with various such labels
and adds them to the attestation record by first
verifying these labels from the device, and then
copying them to the attestation record.

For passing Android compliance tests - CTS and VTS
for Android-13 version, we do not need IMEI, MEID
and Serial Number. Hence, we are not attesting
them in the record.

BUG=b:367450688
TEST=emerge-brya chromeos-base/arc-keymint
TEST=cros_workon_make --board=brya chromeos-base/arc-keymint --test

Change-Id: I7552ae66ff1be80ab7e8b552bfdfeb9312e0a879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5950340
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yao Li <yaohuali@google.com>