libec: load values explicitly from packed structs
In b/374842193#comment4 and comment 5, I wrote a bit about how this code
runs afoul of alignment requirements.

This is pretty subtle undefined behavior; it'd be nice if Clang caught
it. I'll look for an upstream bug (or file one) and post the results
on b/374842193.

BUG=b:374842193
TEST=FEATURES=test emerge-amd64-generic chromeos-base/libec
TEST=with --profile=ubsan. No alignment-related crashes happen.

Change-Id: I6a6b482e60246db8fbe950434b5a66a40c51b174
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5952171
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>