biod: Use expected to return error from StartEnroll/AuthSession
The returned Session object is intended to live for the duration
of the Enrollment/Auth operation/session. It is not intended to
be the return status of the StartEnroll/AuthSession functions.
For returning an error that only pertains to the start mechanism,
we can use expected. This frees up the error handling ability of the
Session object to be able to handle errors at any point in the
Enroll/Auth session.

This effectively reverts https://crrev.com/c/4153043 and reworks
https://crrev.com/c/4159931.

* Change TestStartAuthSessionFailed to TestStartAuthSessionAlreadyEnded
  and TestStartEnrollSessionFailed to
  TestStartEnrollSessionAlreadyEnded, since errors on
  AuthStart/EnrollStart are returned explicitly. An empty Session
  should be handled as success, where the End function has already
  been called.
* Add TestStartEnrollSessionMove to confirm that the base::expected
  wrapped Session move semantic works correctly.

BUG=b:367773410
TEST=cros workon --board=brox start chromeos-base/biod
     cros_run_unit_tests --board=brox --packages chromeos-base/biod
TEST=FEATURES=test emerge-brox chromeos-base/biod

Change-Id: Iabb699cb5dbb8988f035d18ce9b284a555424b3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5955473
Commit-Queue: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>