odml: [CORAL] Cache title generation results
To handle repetitive triggers in short time, where the formed groups are
likely identical, add a cache of group entities and their titles. When
we receive request groups, we first check against all entries of the
cache to see whether any cached group is similar enough to the request
group. If so, we can reuse the title without using the model to generate
one.

BUG=b:374755350
TEST=odml unittests

Change-Id: I90104abb3c842cc846960b1bb3e49832c9d926dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5952290
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>