sepolicy: Add an entry for session_manager from minijail
This is preparation to migrate the user of session_manager
from root to another. We'll use minijail to switch the user
so add an selinux policy for the new one.
We keep the old one for the transition period, and the entry
will be removed after the migration is made sure succeeded.

BUG=b:289228842
TEST=Ran locally. Tryjob

Change-Id: I9c8f049e8c0ee7bfa25ec4d6ac15937e47b37dfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5952292
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: ChromeOS Auto Runner <chromeos-auto-runner@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>