missive: More potential fixes
I think I found one more race that could cause the crash:

in `MissiveImpl` there was the following code:
  storage_module_ = std::move(storage_module_result.value());
  storage_module_->AttachUploadSuccessCb(storage_upload_success_cb_);

It means that `storage_module_` is initialized and made available
*before* `storage_upload_success_cb_` callback is set. In rare cases when storage was found non-empty, `storage_module_` could perform upload
roundtrip *before* `storage_upload_success_cb_` is set, and so it
would attempt to invoke a null callback upon upload completion.

To fix it I removed `AttachUploadSuccessCb` altogether and instead
pass `storage_upload_success_cb_` as a parameter to `StorageModule`
creation, so that by the time it is initialized the callback is
definitely set.

Besides that I made some clean-up, and also made a minor fix: moved
`weak_ptr_factory_` in `ResourceCollectorStorage` to become the last member in the class (previously it was not).

BUG=b:369605450
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=Manual on VM

Change-Id: Ib7f8822bfbc259d95347e89166d717158f23f8aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5958491
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>