ml: update the ica mojom in ml service
As we are switching to a different model(cl/686744071), also update
the mojom in ml service so that we can unblock the work on the
chromium side. The whole feature is behind a flag and not in
production yet.
We also sync the geometry.mojom file from
`ui/gfx/geometry/mojom/geometry.mojom` as it is currently out of
date, and `gfx.mojom.Rect` is not marked stable.

BUG=b:343320265
TEST=emerge and deploy ml

Change-Id: If972300057ab10800506ae0cdd47822bf3a020aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5961376
Tested-by: CJ Huang <chenjih@google.com>
Commit-Queue: CJ Huang <chenjih@google.com>
Reviewed-by: Xinglong (Alan) Luan <alanlxl@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>