init: show logo in developer mode
The ChromeOS logo is displayed during system initialization if:
- the cros_config setting for "firmware splash screen" is explicitly set
  to FALSE (this only applies to Reven devices);
- the display is not yet active.

For non-reven boards, the display can only be active by the time
boot-splash.conf is executing if:
- firmware splash screen feature is enabled
- the device is in developer mode

On devices without the firmware splash screen feature, the logo appears
in normal mode but not in developer mode. This inconsistency confuses
ODMs and QA testers who expect the logo to always appear.

Fix this inconsistency by adding another statement to the if condition.

BUG=b:373519348
TEST=Check logo is shown in developer mode regardless if firmware splash
screen feature is enabled or not.

Change-Id: I56ee92e91b606baa9d936245b4824a20110a2c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5960259
Tested-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Auto-Submit: Paz Zcharya <pazz@google.com>
Commit-Queue: Paz Zcharya <pazz@google.com>