cryptohome: Update the password delay schedule.
After some discussion we have another draft of a proposed delay policy
for pinweaver-passwords. Unlike the prior one, which is literally just a
copy of the PIN policy, this new one is basically the PIN policy but
with an extra 5 free attempts added to it.

This leaves the TODO in place because the proposal is still just a
draft, yet to be finalized at a later date.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: If3d4f4de093785034c4e90cc7c2603a1d9358ffa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5962386
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>