power: Handle lack of cros_ec in temp_logger
On reven devices there is no /dev/cros_ec device. This causes minijail
to fail to start, because the bind-mount cannot be created. Fix by
making the bind-mount conditional in the upstart script.

Also add an early return in get_fans_speed so that it fails without
errors if /dev/cros_ec is missing.

BUG=b:375636855
TEST=Deploy to reven dut, verify temp_logger works again
TEST=Check that the bind-mount still occurs if /dev/cros_ec exists

Change-Id: I6e06950f26c971cd7d6dddd4fa9c2fcc806bce1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5966816
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>