shill: wifi: Enable scan_ssid for initial hidden network connection
This change allows users to connect to a hidden network for the first
time by simply entering the SSID and security configuration.

It achieves this by enabling the `scan_ssid` flag for networks that have
never been connected before. This flag instructs wpa_supplicant to probe
for the specified SSID.

This improvement simplifies the initial connection process for hidden
networks, eliminating the need for users to manually enable the "hidden
network" option in the network configuration.

BUG=b:363899693
TEST=FEATURES=test emerge-rex chromeos-base/shill and run wifi_matfunc
on Rex/Karis/BE200 (with NUC AP)

Additional testing on same setup:
1- Deployed updated Shill on Karis+BE200
2- Connect to hidden network using UI and connection Pass on OOBE and as
logged-in
3- Disable and Re-enable WiFi and check that the hidden network is not
visible in the scan list

Change-Id: I4111fba2866b23521bb35bfa9a8e64d93735364d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5967875
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>