odml: [CORAL] Don't use cached title from other users
Although this should rarely happen in the first place, we need to
prevent reusing cached titles that are generated by other users. Since
the cache reuse logic doesn't require exact match, the cached title
might contain information that doesn't belong to the current inputs. In
the worst case it might contain PII, so we should restrict cache reusing
to the same user.

We record the user info on title cache, instead of clearing all caches
on user logout because post-login coral chips are likely to have similar
inputs to the previous in-session trigger of the same boot cycle. We
don't want to miss the cache opportunity here.

BUG=b:374755350
TEST=odml unittests

Change-Id: I1a173d502368c0c607545d0f51b4747965a2f8d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5970059
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>