login: Move EnsureDirectoryExists to ChromeSetup.
And inlined SetPermissions to the function, because it is the only caller.
Now the check whether its called under root or not by getuid()
in the function is removed. This was used to check whether or not
it's running in unittest. However, once we switch the user
the check won't work any more.

Now, it fchown is always called, but passing -1 is the trick to pass
on unittests. Later, we'll introduce a flag to control the behavior
when needed as a member of ChromeSetup.

BUG=b:289228842
TEST=Tryjob

Change-Id: Id5842ed722edefa1403014c637ff5efdf290e1a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5970934
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>