runtime_probe: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=runtime_probe \
        -n 1 \
        --platform2-commit=2388a3a9862e92131215d237ac1c10695f3fb5a3

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to privacy and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: Icecaff2d6c1244565c5a166263074f5b2bb10d4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5973578
Tested-by: James Hawkins <jhawkins@chromium.org>
Owners-Override: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>