biod: Fix enroll/auth abort on HW error logic
Both RequestEnrollImage and RequestMatch functions start the actual
enrollment and auth attempt in the FPMCU, which will be finalized or
canceled using the returned Session object. The returned session
object will only cleanup the state when the session_weak_factory_ is
included in the object, in the final return line. Thus, returning an
empty Session with only the HW sensor error, the FPMCU will remain in
the previous mode, while the UI will abort.

BUG=b:374858597, b:230478501
TEST=cros_workon_make --board=brox --test chromeos-base/biod

Change-Id: Iaade8e14138c2ea8e4d6cf2c5d490a966a5508da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5955471
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>