cryptohome: Clean up some style violations
There's a few headers that include raw using declarations, which are
generally against the rules. Explicit aliases are allowed, but these are
clearly just for convenience in the header. Fortunately it looks like
nothing was actually depending on these being in the headers so we can
just remove them.

BUG=none
TEST=cq

Change-Id: I4e9c3724d76b01bd626ebab6fa9123e770cfc19b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5981249
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>