vm_tools: concierge: disable pmem rootfs (except for Borealis)
We don't get any benefit from using pmem for the Crostini rootfs, since
it is not shared with any other VMs, and it doesn't take advantage of
DAX. Using virtio-blk on all architectures also makes things more
consistent.

Unfortunately, Borealis expects the root device to be /dev/pmem0 and
fails to boot otherwise, so we keep the pmem code around for that case.

BUG=b:375506171
TEST=deploy vm_host_tools and boot Crostini + ARCVM

Change-Id: I4147bafbc0cc8d18020a56b88114527d5da20f23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5981251
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>