dlp: fix build with linux-headers >= 5.4
`fanotify_reader_thread.cc` includes the definitions for structs which
were absent from the linux headers we shipped so far (4.14), but are now
part of the new uprev target (5.4).

This results in failure to build as those appear to be redefining those
structs. Those definitions will be removed once the linux-headers uprev
is complete, but for now we need to still allow building with the 4.14
headers, so let's just add a preprocessor condition to ensure both cases
work as expected.

BUG=b:189218019
TEST=import CL #5973827 and rebuild image from scratch

Change-Id: I1a57250f3af1eb4d6ce8075c0291f94652f4b301
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5980071
Reviewed-by: Sergey Poromov <poromov@chromium.org>