cros-codecs: Remove MD5 crate dependency
Remove the dependency on the MD5 crate. This crate is problematic to
pull into ChromeOS because MD5 is cryptographically broken, and its
presence in ChromeOS's third party Rust crates may encourage others to
use it in dangerous ways.

We are not using MD5 in a cryptographic capacity, so this concern does
not apply to us. This patch adds a simple implementation of MD5 directly
in cros-codecs for us to keep using existing regression testing
infrastructure.

BUG=b:376859769
TEST=Ran unit tests and simple_test.py, verified no new regressions.

Change-Id: Id4ab73bc520e0b5c151ec7910a61d3e7a46d47ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5994770
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Tested-by: Justin Green <greenjustin@google.com>
Commit-Queue: Justin Green <greenjustin@google.com>