libhwsec-foundation: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=libhwsec-foundation \
        -n 1 \
        --platform2-commit=d0f72ba10e3ba8d5068bb6e3460f9fa9b4f693ac

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to libhwsec and
-n set to 5.

BUG=b:370478040
TEST=CQ

Change-Id: Ifa5429177ea3c9317b461a12c62869813fd66249
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5995094
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>