libtouchraw: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=libtouchraw \
        -n 1 \
        --platform2-commit=7a2342778766b35782b1565f490a5f4251b51fcf

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to libsar and
-n set to 5.

BUG=b:370478040
TEST=CQ

Change-Id: Ib66a49e3237ef54640ae1a983e9b823cb9a73a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6000295
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>