login_manager: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=login_manager \
        -n 1 \
        --platform2-commit=7a2342778766b35782b1565f490a5f4251b51fcf

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to libsar and
-n set to 5.

BUG=b:370478040
TEST=CQ

Change-Id: I037c89da41e1919f4747a852e5ab578ad6c3ec79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6000296
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>