arc: Attest ARCVM serial number
This CL adds the capability to attest the
ARCVM's serial number if it is requested
by an app in the Attestation certificate.

It will look at the serial number provided in the
request, and attest to it if it matches the one
already setup in KeyMint.

BUG=b:372711104
TEST=emerge-brya chromeos-base/arc-keymint
TEST=cros_workon_make --board=brya chromeos-base/arc-keymint --test

Change-Id: Ia5444fc105f768d2abea692c0029add2e4c1311b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6001444
Reviewed-by: Yao Li <yaohuali@google.com>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>