power: intel_prepost_suspend: Add --sighting_check flag
When we have a sighting alert of a platform. All projects with that
platform would be impacted. Usually, different platforms with the
same generation hardward architecture design would have the same
sighting alert.

For example, sighting alert 772439, impacting all projects of 'ADL-P,
ADL-N, TGL, RPL-P, RPL-P Refresh, and TWL. We need to interrupts tests,
requiring, collecting, uploading, and analysis the log. We need to do
it repeatly for every projects stages, every factory build.

To save time, implement the command to verify the sighting alert.
Also this flag would impact the return value, we could use it in any
script we want to check sighting, too.

We could get the result easily when suspend fail:
  'intel_prepost_suspend --sighting_check'
If that is sighting alert case, it returns true and stdout shows
  'CNVi Sighting Alert 772439!'
If not sighting alert, nothing output but return false.

BUG=b:378372771, b:271527450, b:358512465
TEST=Execute 'intel_prepost_suspend --sighting_check' anytime and check
     the response.

Change-Id: I14dac5a4cfc36ba2c100f2e239f4d9d7c80c6547
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5999134
Tested-by: Curtis Chen <curtis.chen@intel.com>
Commit-Queue: Curtis Chen <curtis.chen@intel.com>
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>