shill: net: Service synchronizes counters with attached Network
This patch changes how a Service synchronizes its traffic counters
from patchpanel. Instead of querying patchpanel directly, it delegates
the query to its attached Network. The raw counter snapshot is
reinitialized when a Network is attached, and the aggregated counters
are updated when a Network is detached.

It is also possible to request traffic counters from a Service even if
no Network is attached. This returns the currently aggregated counters
loaded from storage or from the in-memory state after the last
disconnection.

The connection state of the attached Network is not taken into account.
This is already implicitly included in the raw traffic counters (i.e the
raw traffic counters are not changing when the Network is disconnected).

Any other secondary Network for that Service or its Device are also not
taken into account.

Finally, the Service will not load into memory empty counters for a
given source if traffic for that source has never been recorded for that
Service.

BUG=b:273741930
BUG=b:335299463
TEST=emerge-$BOARD shill, manually checked counter increasing from DBus
Service.RequestTrafficCounters.

Change-Id: I2efcd903138cdd3e3431b5e39a36416f8d43ad26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6012144
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>