u2fd: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=u2fd \
        -n 1 \
        --platform2-commit=e3cd3e680636d5aa82fb2716ba465ee54b1a328c

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` unset and
-n set to 5.

BUG=b:370478040
TEST=CQ

Change-Id: I251ced6738f29cd84e7049bc06c2aa71d727fd7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6001182
Tested-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Owners-Override: James Hawkins <jhawkins@chromium.org>