ureadahead-diff: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=ureadahead-diff \
        -n 1 \
        --platform2-commit=2d2489747ae9b9b6feb2771792a92bf5a3990ce2

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to ureadahead-diff and
-n set to 5.

BUG=b:370478040
TEST=CQ

Change-Id: Ib6ffcc263c03d806449012949d7295b564292f5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6002600
Tested-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Owners-Override: James Hawkins <jhawkins@chromium.org>