libhwsec-foundation: Add a const overload for StatusOr ->
Minor improvement, but I ran into this in a couple places where I have a
const StatusOr and it's a bit annoying to not be able to use -> to call
(const) methods on the contained value.

This makes xxx->call() equivalent to xxx.value().call() in the const
case, which is already true for non-const StatusOr objects.

BUG=none
TEST=unit tests

Change-Id: I3d39709ed332cfd74076955b6f056ba1bf2da86c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6003886
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>