crash: Handle A1 NoC errors properly
Sometimes we see NoC errors that look like

 A1NOC ERROR: ERRLOG0_LOW

but the regex fails because it doesn't expect the type of NoC to be
directly attached to the letters "NOC". Update the regex so this works
correctly by making the underscore optional and also allow numeric
values in the NoC name.

BUG=None
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I6a8f2616e94a674f396f3314f5620ecd558a217e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6006236
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>