libchromeos-rs: Convert TempBackedMount to Mount
Create a MountPoint enum to abstract away the details of how we're
holding our mount, in preparation for adding a `Mount` that doesn't
manage a tempdir.

I recently landed a similar change with the names going in the opposite
direction, but this better way was brought up in code review.

BUG=b:356344778
TEST=cros workon -breven start \
               dev-rust/libchromeos flexor os_install_service
     cros build-packages -breven \
               dev-rust/libchromeos flexor os_install_service
     cros_sdk FEATURES=test -- emerge-reven \
               dev-rust/libchromeos flexor os_install_service
TEST=CARGO_TARGET_X86_64_UNKNOWN_LINUX_GNU_RUNNER='sudo -E' \
               cargo test -- --ignored

Change-Id: I1dcd76f4d22832e248744e9bb1c87fab4ccdf391
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6011272
Commit-Queue: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>