libbrillo: Fix unsafety in operator~ in enum_flags.h
Valid values for an enum must be the smallest and largest values of a
bit-field that can store the minimum and maximum values of the enum.

However, enum_flags.h's `operator~` returned values for enums which are
outside of the value. This causes ASAN errors, see attached bug.

Fix this by returning the underlying type from `operator~`, and allowing
`operator&` to take in the underlying type too. This fixes the
aforementioned ASAN errors.

BUG=b:329990991
TEST=CQ

Change-Id: I6a6a2100c2f7f77f2f3a97ebda7a016b26b6adde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6014058
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Auto-Submit: Michael Cui <mlcui@google.com>
Tested-by: Michael Cui <mlcui@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>