cros-codecs: Backport 080717f23d2f3ff0dc34ac79ce02fe9ca12525f1
Original CL author: Alexandre Courbot <acourbot@chromium.org>

Original commit message:

codec/h264: separate DPB entry reference and frame

The DPB serves two purposes: as a way to reorder frames in display
order, and as a way to mark frames that are still used as reference.

Currently both purposes are served by the same `handle` member, but
separating it into dedicated members brings benefits: the type of the
decoded frame could be different from the one used for reference, and
the decoded frame could be take()n away from an Option when the frame is
bumped, making sure it cannot be emitted a second time.

IOW, this should allow the removal of the Rcs to reference frames in the
DPB, once a proper reference frame type is introduced.

BUG=b:378665810
TEST=video.PlatformDecoding.cros_codecs*

Change-Id: I13de6b523bf4ebe9577853c9522fdd61a3763e4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6015090
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Justin Green <greenjustin@google.com>
Tested-by: Justin Green <greenjustin@google.com>