installer: Add Environment type for passing to sh
Add an Environment type to represent key-value pairs for passing to
Command::envs. This type holds (String, OSString), rather than
(String, String) to reduce some type conversion, e.g. when we have a
Path we can avoid converting it to a String (and the error handling
that goes with that conversion).

BUG=b:356344778
TEST=cargo test
TEST=cargo xtask test-install chromiumos_test_image.bin

Change-Id: I032297dc1dd24a1fa7997ff59d712774c5d1c86f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6011607
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>