cros-codecs: Backport 9979b2ba3c622882e7f586754b6dea8bc60b4d61
Original CL author: Alexandre Courbot <acourbot@chromium.org>

Original commit message:

decoder/stateless/h264: remove last_field

This field is only used with interlaced streams, and in case the DPB is
full when we try to add the first field of a frame to it.

It has been confirmed with Daniel Almeida (who initially wrote this
code) that this field was added as an extra safety to mimic something
GStreamer does, but is not warranted by the specification itself.

In practice, none of our unit tests or Fluster streams ever exercise a
path where last_field is set, and its presence seriously complicates the
code. Let's remove it.

BUG=b:378665810
TEST=video.PlatformDecoding.cros_codecs*

Change-Id: I34199368abdd7b980e533d4e5f108765ccddd583
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6015094
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Tested-by: Justin Green <greenjustin@google.com>
Commit-Queue: Justin Green <greenjustin@google.com>