init: startup: Set exception when needed
After dev clobber, /var/lib is removed. It is added later, but the
symlink exemption is not preserved.
Readd it after clobbering, if needed.

Also if we clobber on the first install, /var/ and dev_image will be
replaced during install, so we have to resend the symlink execptions for
these directories.

BUG=b:373722016
TEST=Check after dev clobber the traversal error is gone:
```
Blocked symlink traversal for path fe:1:/var/lib/timezone/localtime
```

Change-Id: I763310b27d56b89af1eb0e98b719a8bf3ac93120
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6015905
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>