lorgnette: sane_device_impl: Refactor GetX/YRange to GetMaxWidth/Height
The clients of GetX/YRange() only use the size field of the
OptionRange struct to determine the maximum scannable width/height.

Change the GetX/YRange() functions to GetMaxWidth/Height() to be more
explicit about the return values and how they are to be used.

BUG=b:377679004
TEST=FEATURES="test" emerge-${BOARD} lorgnette

Change-Id: I3036fa58e6e03c85064590e66e491df9e641c201
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6026638
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Aaron Massey <aaronmassey@google.com>