libec: return the correct number of dead pixels
When other error flags are set, the number of dead pixels returned was
not correct. Use the FP_ERROR_DEAD_PIXELS macro to extract the number of
dead pixels from the error variable. Generalize the unit tests to
include additional error flags to ensure that the correct number of dead
pixels are extracted.

BUG=b:379143410
TEST=cros_sdk cros_workon_make --board=hatch libec --test

Change-Id: I7e7fc0a01039032479f4302bb2364d0ebea88b8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6032216
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Tested-by: Firas Sammoura <fsammoura@google.com>