cryptohome: Flip PIN migration flag to default-enabled
This flag is now enabled in production and so we should flip the
default. A few tests needed to be updated to inject a delay schedule
value but otherwise for the most part the tests had already been
properly updated to not depend on the flag default.

BUG=b:379037653
TEST=unit tests, cq

Change-Id: I83e1201efc2c3a78a3ad1ddf2f719d89d0218e2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6033093
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>