diagnostics: Allow touchpads connected via usb in TouchpadFetcher
Initially it was thought that internal touchpads would not be connected
via usb and this was a good enough check for if a touchpad is internal
or external. As it turns out, some Macbooks have internal touchpads
which are connected via the usb bus, so they were not reporting touchpad
information as expected.

This change removes the check for being on the USB bus to allow for
those touchpads to be correctly reported.

BUG=b:379872627
TEST=`USE="-cros-debug" emerge-reven diagnostics`
TEST=`(on dut) cros-health-tool telem --category=input`
TEST=`FEATURES=test emerge-reven diagnostics`

Change-Id: I3222628a2cea978efc2986f6e7497361dd20159a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6039173
Reviewed-by: Ethan Cheng <yycheng@google.com>
Commit-Queue: Jonathan Denose <jdenose@google.com>
Tested-by: Jonathan Denose <jdenose@google.com>