login: Rename SystemUtils variables
Conceptually, it should be a global singleton, but it was referred
by variation of names depending on places, such as system, utils,
system_utils, which is syntax noises on reading the code.

This CL just renames all SystemUtils variables to system_utils
(or system_utils_ for member variables).

BUG=b:289228842
TEST=Tryjob

Change-Id: I5edb1453e75671c04422593216b0db800f402674
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6039400
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>