fpstudy: Migrate detailed BPET requirements to standalone doc
There is a lot of complexity and requirements about what we expect
vendors to provide in a biometric performance evaluation tool, so let's
move this into its own document.

Rename the PET to BPET, to separate it any latency/timing performance
evaluation, per Firas suggestion. If we incorporate timing and latency
evaluation into this tool in the future, we can change back to PET.

Rewrite the introduction to BPET similar to Firas's opener description.

BUG=b:373433931
TEST=mdformat --compatibility --w ANALYSIS.md BPET.md
TEST=Inspect with Gitles

Change-Id: I5ad3f6eb87a65e976600b7832d6bd52dd61961ae

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6044288
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>