fpstudy: Move the collection tool into sub directory
This makes it more clear what source and components contribute to the
"collection" part. We should now reffer to this part of the fingerprint
study tool, as the fingerprint collection tool or fpstudy collection
tool.

That means that we now have the following components to an FP study:
* Fingerprint Study Image Builder
* Fingerprint Study Collection Tool
* Biometric Performance Evaluation Tool (BPET)
* Fingerprint Study Analysis Tool

BUG=b:339240874
TEST=cros workon --board=brox start fingerprint_study
     cros build-packages --board=brox fingerprint_study
     cd ~/chromiumos/out/build/brox
     ls -al opt/google/fingerprint_study/*

Cq-Depend: chromium:6046289
Change-Id: I7d8e058396343d7019829b4123bd0076d0ebc482

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6044505
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>