power: Change KBBL timeout policy
Currently, if the user sets the keyboard backlight value manually or if
it is restored by Chrome after booting, the timeout for the keyboard
backlight is disabled and it is instead tied to the display timeout.
This change makes it so that the normal timeout is always used.

Update unit tests to remove tests for the old policy which disabled the
timeout; instead test that the timeout is always applied.

Documentation will be updated in a later CL.

BUG=b:377189886, b:379701507
TEST=update power_manager, check that the backlight timeout works
TEST=cros_run_unit_tests --board=brya --packages "power_manager"

Change-Id: Icdf77b052dae6a9e3a0e6c68e7c944d433e09560
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6088792
Commit-Queue: Eric Herrmann <eherrmann@chromium.org>
Auto-Submit: Eric Herrmann <eherrmann@chromium.org>
Tested-by: Eric Herrmann <eherrmann@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>