cros-codecs: Return error for buffer allocation failure
The upstream commit: b9a7177
decoder/stateless/h264: let the backend decide which resources it needs

changed the allocation of buffers so that buffer recycling is
only done when an output buffer fails allocation.

The backpressure from the output queue allocation failing results
in the recycling of buffers. Previously the allocation failure
was unexpected and produced a panic because the failure was not
handled gracefully.

Returning an error when this allocation fails now allows the
buffers to be returned to the queue for subsequent allocation
requests.

BUG=b:380903992
TEST=ccdec-v4l2-stateless-chromeos ${CLIP} --input-format h264

Change-Id: Ib1bc4b549b606af079ac604f1ccec14ce44d586b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6068410
Tested-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Steve Cho <stevecho@chromium.org>