cryptohome: Rename unittest.cc -> test.cc
This is actually fixing a TODO/lint error, albeit a very noisy fix. But
now that code change volume is done it actually is a fairly good time to
do it while minimizing merge conflicts.

It is a bit rough on source control history, but that's unfortunately
just the way that renames are.

BUG=none
TEST=unit tests (they still run with the new filenames!)

Change-Id: I04cc7b6186b67472ca4730a037f1feac0bd4e4d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6072182
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>