fpstudy: Add sourceme.bash to enable Python venv and cmd tool search
This adds the sourceme.bash script to setup your local bash shell
for working with collection-tool and analysis-tool. Its main function
is to activate the Python venv and add the analysis-tool dir to the
bash command search PATH. These two things make it easier to develop
for collection-tool and analysis-tool.

BUG=b:339240874
TEST=./study/python-venv-setup.sh
     . ./study/sourceme.bash
     which python3
     # Should report the python3 in the .venv/bin directory.
TEST=./study/python-venv-setup.sh
     . ./study/sourceme.bash
     test_case_test.py

Change-Id: I386607fe2d9ebe8f6db0a93cbd5adf1a6e7e9e8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6072522
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>