fpstudy: Add check in python-venv-setup.sh for active venv
It is a problem to run the python-venv-setup.sh script when the
venv that we are about to remove and rebuild is currently active.
This would mean that we are invoking the scoped python3 command inside
the venv directory that we are about to remove.

BUG=b:339240874
TEST=./study/python-venv-setup.sh
     . ./study/sourceme.bash
     ./study/python-venv-setup.sh
     # Should fail
     deactivate
     ./study/python-venv-setup.sh
     # Should succeed

Change-Id: I89afe907863db8efb7c56f0e19f2a35dfb5d542b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6072525
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>