screen-capture-utils: Read the COLOR_ENCODING/COLOR_RANGE plane props
This CL adds plumbing to read the COLOR_ENCODING and COLOR_RANGE DRM
plane properties. The newly read properties are not used in this CL.
They're used in a follow-up CL to ensure YUV-to-RGB conversion is made
using the correct computations. This is useful because the
video.Contents.* Tast tests (which use the `screenshot` utility) are now
more deliberate about color spaces, so we need improved color
correctness.

Following Chrome's lead [1], these properties are made
optional. That said, also following Chrome's lead [2], consistency is
also enforced: either both are present or both are absent.

To handle the plumbling of both the existing required properties and the
new optional/enum properties, the PopulatePlaneConfiguration() function
is refactored slightly.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ui/ozone/platform/drm/gpu/hardware_display_plane.h;l=80-87;drc=3fa7db819d49ef970fc3642f50c90ef09f6a4e8a
[2] https://source.chromium.org/chromium/chromium/src/+/main:ui/ozone/platform/drm/gpu/hardware_display_plane_atomic.cc;l=90-93;drc=2429a70f4f75a5b0d6483557c8ca261b42403618

BUG=b:371017664
TEST=Run `screenshot out.png` on a brya device

Change-Id: I3367acfd36e423bf8c1c7e8d5db5f33125599cc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6072344
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Commit-Queue: Andres Calderon Jaramillo <andrescj@google.com>
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>