print_tools: Add filtering options to printer_diag
When investigating behavior of attributes or groups of attributes across
a set of printers, printer_diag can be used together with ippfind to
quickly send a bunch of Get-Printer-Attributes requests.  However,
because it always requests a full dump of all the attributes, the output
usually requires a significant amount of post-processing to extract the
desired attributes.  To make this easier, add three new command-line
options that work together:

--requested_attributes: Comma-separated list of attributes to send as
requested-attributes instead of the default.  This can be used to test
printer behavior with invalid combinations or just to speed up the query
when a specific subset is needed.

--document_format: MIME type to send as document-format.  Since some
attributes vary by document-format, this allows matching the format that
will be sent by ChromeOS during a real request.

--filter: If specified, only emit attributes containing this string in
their name.  This makes it easy to fetch groups of related attributes
without exhaustively enumerating the names up front.

Example output for --filter=scaling --document_format=image/pwg-raster --requested_attributes=all:
{
   "response": {
      "printer-attributes": {
         "print-scaling-default": {
            "type": "keyword",
            "value": "auto"
         },
         "print-scaling-supported": {
            "type": "keyword",
            "value": [ "auto", "auto-fit", "fill", "fit", "none" ]
         }
      }
   },
   "status": "successful-ok"
}

BUG=None
TEST=Manual queries with lab printers

Change-Id: I19fbe9fa0c9b027bed56bfedac3096c06a605990
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6085755
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>