print_tools: Introduce simplified printer_diag output
The new --simplified flag still emits JSON, but the structure is
simplified to make it easier for human consumption.  The main
attribute groups are returned as top-level keys in the response, and
individual attributes are returned as key-value pairs instead of
sub-objects.

When combined with the previous filter and attribute arguments, this can
be used from ippfind to quickly check sets of attributes across all
local printers.

Example output with --filter=marker --simplified=-:
{
   "printer-attributes": {
      "marker-colors": [ "#000000", "#00FFFF", "#FF00FF", "#FFFF00", "#7FFFFF", "#FF7FFF" ],
      "marker-high-levels": [ 100, 100, 100, 100, 100, 100 ],
      "marker-levels": [ 81, 86, 87, 88, 85, 87 ],
      "marker-low-levels": [ 15, 15, 15, 15, 15, 15 ],
      "marker-names": [ "Black ink", "Cyan ink", "Magenta ink", "Yellow ink", "Light Cyan ink", "Light Magenta ink" ],
      "marker-types": [ "inkCartridge", "inkCartridge", "inkCartridge", "inkCartridge", "inkCartridge", "inkCartridge" ]
   },
   "status": "successful-ok"
}

This format does not reproduce the original IPP frame with full
fidelity, so it does not replace the original JSON output.

BUG=None
TEST=Manual review from lab printer responses

Change-Id: I0fbd87a5ca5d80df1e0168fc3765b1d62df99777
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6085756
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>